requirements.txt
click
dcor
numpy
scikits.bootstrap
scipy
sklearn